update: handleDeleteUpdate

+++ src/pages/Registries/List/index.js
@@ -50,13 +50,16 @@ export default function ListRegistries() {
   }
 
   function handleDelete(registryId) {
-    dispatch(deleteRequest(registryId));
+    // eslint-disable-next-line
+    if (window.confirm(`Excluir plano?`)) {
+      dispatch(deleteRequest(registryId));
+    }
   }
 
   return (
     <ListLayout>
       <>
-        <ListHeader pageName="matrículas" dest="/registry/store" />
+        <ListHeader pageName="matrículas" dest="/registry/new" />
         <RegistryTable
           registries={registries}
           handleDelete={handleDelete}
@@ -66,47 +69,3 @@ export default function ListRegistries() {
     </ListLayout>
   );
 }
-
-/* <ListLayout>
-      <>
-        <table>
-          <thead>
-            <tr>
-              <th className="left">ALUNO</th>
-              <th className="center">PLANO</th>
-              <th className="center">INÍCIO</th>
-              <th className="center">TÉRMINO</th>
-              <th className="center">ATIVA</th>
-              <th> </th>
-            </tr>
-          </thead>
-          <tbody>
-            {registries.map(registry => (
-              <tr key={registry.student_id}>
-                <td className="left">{registry.student.name}</td>
-                <td className="center">{registry.plan.title}</td>
-                <td className="center">{registry.startDate}</td>
-                <td className="center">{registry.endDate}</td>
-                <td className="center">
-                  <IoMdCheckmarkCircle
-                    size={20}
-                    color={registry.active ? "#42cb59" : "#ddd"}
-                  />
-                </td>
-                <td>
-                  <button type="button" onClick={() => handleEdit(registry)}>
-                    <Link to="/edit/registries">editar</Link>
-                  </button>
-                  <button
-                    type="button"
-                    onClick={() => handleDelete(registry.student_id)}
-                  >
-                    apagar
-                  </button>
-                </td>
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </>
-    </ListLayout> */
